[Major upgrade] Process the dimensions provided by prodigy_xy plugin.
Now the prodigy_xy plugin returns the data with
the original names and values of the second and third
dimensions (non energy and rotation).

The plugin IF_UMCS needs to convert non energy channels
to angle values depending on the lens that was used.

The conversion between nonenergy value provided in mm
is done using specific deg/mm scaling or magnification.